+++ packages/adapter-utils/src/utils.ts
@@ -471,7 +471,6 @@ export const getPath = (
   rootElement: Element,
   fullElemID: ElemID
 ): string[] | undefined => {
-  if (rootElement.elemID.isEqual(fullElemID)) return []
   const { parent, path } = fullElemID.createTopLevelParentID()
   if (!_.isEqual(parent, rootElement.elemID)) return undefined
   if (_.isEmpty(path)) return []
